requirements.txt
Flask
python-pptx
requests
boto3
